+++ src/modules/genres/genres.controller.ts
@@ -34,7 +34,7 @@ export class GenresController {
   @MinimumRole(Role.GUEST)
   @ApiOkResponse({ type: () => Genre, isArray: true })
   async getGenres(): Promise<Genre[]> {
-    const genres = await this.genreRepository.find();
+    const genres = await this.genreRepository.find({ relations: ["games"] });
     genres.sort((a, b) => b.games?.length - a.games?.length);
     return genres;
   }