+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # GameVault Backend Server Changelog
 
-## Upcoming
+## 9.0.4
 
 Recommended Gamevault App Version: `v1.7.3`
 
 ### Changes
 
+- Fixed Server not finding users with only deleted progresses
 - Fixed `GET /api/genres` not returning list of genres sorted descending by count of games
 
 ## 9.0.3